I want both timeline and calander switch  between
